+++ yash-syntax/src/parser/case.rs
@@ -31,10 +31,9 @@ use crate::syntax::CompoundCommand;
 impl Parser<'_, '_> {
     /// Parses a case item.
     ///
-    /// Does not parse the optional trailing double semicolon.
-    ///
-    /// Returns `None` if the next token is `esac`.
-    pub async fn case_item(&mut self) -> Result<Option<CaseItem>> {
+    /// If the next token is `esac`, returns `None` without consuming it.
+    /// Otherwise, returns a case item and whether there may be a next item.
+    pub async fn case_item(&mut self) -> Result<Option<(CaseItem, bool)>> {
         fn pattern_error_cause(token_id: TokenId) -> SyntaxError {
             match token_id {
                 Token(_) => unreachable!(),
@@ -105,11 +104,19 @@ impl Parser<'_, '_> {
         let body = self.maybe_compound_list_boxed().await?;
         let continuation = CaseContinuation::default();
 
-        Ok(Some(CaseItem {
-            patterns,
-            body,
-            continuation,
-        }))
+        let continued = self.peek_token().await?.id == Operator(SemicolonSemicolon);
+        if continued {
+            self.take_token_raw().await?;
+        }
+
+        Ok(Some((
+            CaseItem {
+                patterns,
+                body,
+                continuation,
+            },
+            continued,
+        )))
     }
 
     /// Parses a case conditional construct.
@@ -156,13 +163,11 @@ impl Parser<'_, '_> {
         }
 
         let mut items = Vec::new();
-        while let Some(item) = self.case_item().await? {
+        while let Some((item, continued)) = self.case_item().await? {
             items.push(item);
-
-            if self.peek_token().await?.id != Operator(SemicolonSemicolon) {
+            if !continued {
                 break;
             }
-            self.take_token_raw().await?;
         }
 
         let close = self.take_token_raw().await?;
@@ -220,10 +225,12 @@ mod tests {
         let mut lexer = Lexer::from_memory("foo)", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 1);
         assert_eq!(item.patterns[0].to_string(), "foo");
         assert_eq!(item.body.0, []);
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(!continued);
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
         assert_eq!(next.id, EndOfInput);
@@ -234,10 +241,12 @@ mod tests {
         let mut lexer = Lexer::from_memory("(foo)", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 1);
         assert_eq!(item.patterns[0].to_string(), "foo");
         assert_eq!(item.body.0, []);
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(!continued);
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
         assert_eq!(next.id, EndOfInput);
@@ -248,12 +257,14 @@ mod tests {
         let mut lexer = Lexer::from_memory("1 | esac | $three)", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 3);
         assert_eq!(item.patterns[0].to_string(), "1");
         assert_eq!(item.patterns[1].to_string(), "esac");
         assert_eq!(item.patterns[2].to_string(), "$three");
         assert_eq!(item.body.0, []);
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(!continued);
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
         assert_eq!(next.id, EndOfInput);
@@ -264,12 +275,14 @@ mod tests {
         let mut lexer = Lexer::from_memory("foo)\necho ok\n:&\n", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 1);
         assert_eq!(item.patterns[0].to_string(), "foo");
         assert_eq!(item.body.0.len(), 2);
         assert_eq!(item.body.0[0].to_string(), "echo ok");
         assert_eq!(item.body.0[1].to_string(), ":&");
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(!continued);
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
         assert_eq!(next.id, EndOfInput);
@@ -280,13 +293,15 @@ mod tests {
         let mut lexer = Lexer::from_memory("foo);;", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 1);
         assert_eq!(item.patterns[0].to_string(), "foo");
         assert_eq!(item.body.0, []);
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(continued);
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
-        assert_eq!(next.id, Operator(SemicolonSemicolon));
+        assert_eq!(next.id, EndOfInput);
     }
 
     #[test]
@@ -294,14 +309,16 @@ mod tests {
         let mut lexer = Lexer::from_memory("foo):;\n;;", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 1);
         assert_eq!(item.patterns[0].to_string(), "foo");
         assert_eq!(item.body.0.len(), 1);
         assert_eq!(item.body.0[0].to_string(), ":");
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(continued);
 
         let next = parser.peek_token().now_or_never().unwrap().unwrap();
-        assert_eq!(next.id, Operator(SemicolonSemicolon));
+        assert_eq!(next.id, EndOfInput);
     }
 
     #[test]
@@ -322,10 +339,12 @@ mod tests {
         let mut lexer = Lexer::from_memory("(esac)", Source::Unknown);
         let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
-        let item = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
+        let (item, continued) = parser.case_item().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(item.patterns.len(), 1);
         assert_eq!(item.patterns[0].to_string(), "esac");
         assert_eq!(item.body.0, []);
+        assert_eq!(item.continuation, CaseContinuation::Break);
+        assert!(!continued);
     }
 
     #[test]